Unify data processing
